Created variables, tested access to html elements
Initialized variables for score. Created variables for each html
element I will need to access.

Tested that the variables were in fact accessing the html elements
by adding an event listener for click on the rock button.

When the test failed I found out the script needed to be moved
the end of the body tag to ensure the browser has created the
full DOM by the time the script is read.