Built all the functionality
Added a game function called when one of the buttons is clicked.
Added a determine winner function to compare the user's choice
with a randomly generated computer choice. Built in a play again
button to be displayed when one of the players hits 5 wins, which
calls a reset game function.

Also moved the randomness test to the footer of the page with a
message to the user for how to use it. Because Bob is surprisingly
good.